jitu commit

+++ config/src/main/java/com/gl/ceir/config/controller/ConsignmentController.java
@@ -137,9 +137,15 @@ public class ConsignmentController {
 	public GenricResponse deleteConsigment(@RequestBody ConsignmentUpdateRequest consignmentUpdateRequest) {
 
 		logger.info("Consignment Withdraw Request ="+consignmentUpdateRequest);
-
-		GenricResponse genricResponse =	consignmentServiceImpl.deleteConsigmentInfo(consignmentUpdateRequest);
+		GenricResponse genricResponse=null;
+		if(consignmentServiceImpl.updatePendingApproval(consignmentUpdateRequest)) {	
+			genricResponse =	consignmentServiceImpl.deleteConsigmentInfo(consignmentUpdateRequest);
 		logger.info("Response of Delete Request="+genricResponse);
+		}
+		else {
+			new GenricResponse(1, "Error during update status before deleting", consignmentUpdateRequest.getTxnId());
+		}
+	
 
 		return genricResponse;
